# starter/gin/docs/swagger.yaml
basePath: /api
definitions:
...
  product.CreateProductDTO:
    properties:
      code:
        maxLength: 50
        minLength: 2
        type: string
      description:
        maxLength: 500
        type: string
      layout_ref:
        maxLength: 100
        type: string
      max_seats:
        minimum: 0
        type: integer
      name:
        maxLength: 100
        minLength: 2
        type: string
      price:
        minimum: 0
        type: number
      status:
        enum:
          - active
          - inactive
        type: string
      trial_days:
        minimum: 0
        type: integer
      type:
        enum:
          - personal
          - team
        type: string
    required:
      - code
      - name
      - type
    type: object
...
  product.ProductDTO:
    properties:
      code:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      layout_ref:
        type: string
      max_seats:
        type: integer
      name:
        type: string
      price:
        type: number
      status:
        type: string
      trial_days:
        type: integer
      type:
        type: string
      updated_at:
        type: string
    type: object
  product.UpdateProductDTO:
    properties:
      code:
        maxLength: 50
        minLength: 2
        type: string
      description:
        maxLength: 500
        type: string
      layout_ref:
        maxLength: 100
        type: string
      max_seats:
        minimum: 0
        type: integer
      name:
        maxLength: 100
        minLength: 2
        type: string
      price:
        minimum: 0
        type: number
      status:
        enum:
          - active
          - inactive
        type: string
      trial_days:
        minimum: 0
        type: integer
      type:
        enum:
          - personal
          - team
        type: string
    type: object
...
  response.DataResponse-product_ProductDTO:
    properties:
      code:
        description: HTTP 状态码
        type: integer
      data:
        allOf:
          - $ref: "#/definitions/product.ProductDTO"
        description: 响应数据
      error:
        description: 错误详情（仅失败时）
      message:
        description: 消息描述
        type: string
    type: object
...
  response.ErrorDetail:
    properties:
      code:
        description: 业务错误码（小写下划线）
        type: string
      details:
        description: 额外详情（如验证错误列表）
      message:
        description: 错误消息
        type: string
    type: object
  response.ErrorResponse:
    properties:
      error:
        $ref: "#/definitions/response.ErrorDetail"
    type: object
  response.MessageResponse:
    description: 纯消息响应格式
    properties:
      code:
        description: HTTP 状态码
        type: integer
      message:
        description: 消息描述
        type: string
    type: object
...
  response.PagedResponse-product_ProductDTO:
    properties:
      code:
        description: HTTP 状态码
        type: integer
      data:
        description: 数据列表
        items:
          $ref: "#/definitions/product.ProductDTO"
        type: array
      message:
        description: 消息描述
        type: string
      meta:
        allOf:
          - $ref: "#/definitions/response.PaginationMeta"
        description: 分页信息
    type: object
...
  response.PaginationMeta:
    properties:
      has_more:
        description: 是否有下一页
        type: boolean
      page:
        description: 当前页码
        type: integer
      per_page:
        description: 每页数量
        type: integer
      total:
        description: 总记录数
        type: integer
      total_pages:
        description: 总页数
        type: integer
      warning:
        description: 页码越界警告
        type: string
    type: object
...
host: localhost:40012
info:
  contact:
    email: lwmacct@icloud.com
    name: API Support
    url: https://github.com/lwmacct/260101-go-pkg-ddd/issues
  description: 基于 DDD+CQRS 架构的 Go Web 应用 API 文档
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/lwmacct/260101-go-pkg-ddd
  title: Go DDD Template API
  version: 1.0.3
paths:
...
  /api/admin/products:
    get:
      consumes:
        - application/json
      description: 分页获取产品列表
      parameters:
        - default: 20
          description: Limit 每页数量，默认 20，最大 1000
          in: query
          maximum: 1000
          minimum: 1
          name: limit
          type: integer
        - default: 1
          description: Page 页码，从 1 开始
          in: query
          minimum: 1
          name: page
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: 产品列表
          schema:
            $ref: "#/definitions/response.PagedResponse-product_ProductDTO"
        "401":
          description: 未授权
          schema:
            $ref: "#/definitions/response.ErrorResponse"
        "403":
          description: 权限不足
          schema:
            $ref: "#/definitions/response.ErrorResponse"
        "500":
          description: 服务器内部错误
          schema:
            $ref: "#/definitions/response.ErrorResponse"
      security:
        - BearerAuth: []
      summary: 产品列表
      tags:
        - Admin - Product Management
    post:
      consumes:
        - application/json
      description: 系统管理员创建可订阅的产品
      parameters:
        - description: 产品信息
          in: body
          name: request
          required: true
          schema:
            $ref: "#/definitions/product.CreateProductDTO"
      produces:
        - application/json
      responses:
        "201":
          description: 产品创建成功
          schema:
            $ref: "#/definitions/response.DataResponse-product_ProductDTO"
        "400":
          description: 参数错误或产品名已存在
          schema:
            $ref: "#/definitions/response.ErrorResponse"
        "401":
          description: 未授权
          schema:
            $ref: "#/definitions/response.ErrorResponse"
        "403":
          description: 权限不足
          schema:
            $ref: "#/definitions/response.ErrorResponse"
        "500":
          description: 服务器内部错误
          schema:
            $ref: "#/definitions/response.ErrorResponse"
      security:
        - BearerAuth: []
      summary: 创建产品
      tags:
        - Admin - Product Management
  /api/admin/products/{id}:
    delete:
      consumes:
        - application/json
      description: 删除产品（软删除）
      parameters:
        - description: 产品ID
          in: path
          name: id
          required: true
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: "#/definitions/response.MessageResponse"
        "401":
          description: 未授权
          schema:
            $ref: "#/definitions/response.ErrorResponse"
        "403":
          description: 权限不足
          schema:
            $ref: "#/definitions/response.ErrorResponse"
        "404":
          description: 产品不存在
          schema:
            $ref: "#/definitions/response.ErrorResponse"
        "500":
          description: 服务器内部错误
          schema:
            $ref: "#/definitions/response.ErrorResponse"
      security:
        - BearerAuth: []
      summary: 删除产品
      tags:
        - Admin - Product Management
    get:
      consumes:
        - application/json
      description: 获取产品详细信息
      parameters:
        - description: 产品ID
          in: path
          name: id
          required: true
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: 产品详情
          schema:
            $ref: "#/definitions/response.DataResponse-product_ProductDTO"
        "401":
          description: 未授权
          schema:
            $ref: "#/definitions/response.ErrorResponse"
        "403":
          description: 权限不足
          schema:
            $ref: "#/definitions/response.ErrorResponse"
        "404":
          description: 产品不存在
          schema:
            $ref: "#/definitions/response.ErrorResponse"
        "500":
          description: 服务器内部错误
          schema:
            $ref: "#/definitions/response.ErrorResponse"
      security:
        - BearerAuth: []
      summary: 产品详情
      tags:
        - Admin - Product Management
    put:
      consumes:
        - application/json
      description: 更新产品信息
      parameters:
        - description: 产品ID
          in: path
          name: id
          required: true
          type: integer
        - description: 更新信息
          in: body
          name: request
          required: true
          schema:
            $ref: "#/definitions/product.UpdateProductDTO"
      produces:
        - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: "#/definitions/response.DataResponse-product_ProductDTO"
        "400":
          description: 参数错误
          schema:
            $ref: "#/definitions/response.ErrorResponse"
        "401":
          description: 未授权
          schema:
            $ref: "#/definitions/response.ErrorResponse"
        "403":
          description: 权限不足
          schema:
            $ref: "#/definitions/response.ErrorResponse"
        "404":
          description: 产品不存在
          schema:
            $ref: "#/definitions/response.ErrorResponse"
        "409":
          description: 产品名称已存在
          schema:
            $ref: "#/definitions/response.ErrorResponse"
        "500":
          description: 服务器内部错误
          schema:
            $ref: "#/definitions/response.ErrorResponse"
      security:
        - BearerAuth: []
      summary: 更新产品
      tags:
        - Admin - Product Management
...
securityDefinitions:
  BearerAuth:
    description: 'Bearer token authentication. Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"